Moved util methods to separate classes
